+++ src/form/b-form/CHANGELOG.md
@@ -26,9 +26,10 @@ Changelog
 
 * Added `getElValueToSubmit`
 * Added the `submitEnd` event
-* Added `ValidationError.component` and `ValidationError.error`
+* Added `ValidationError.name/component/details`
 * Added `getValues`
 * Added `toggleControls`
+* Now `ValidationError` is a class
 * Improved submit events
 
 #### :house: Internal